Move general test utils to a separate crate
