+++ crates/papyrus_gateway/Cargo.toml
@@ -25,5 +25,6 @@ assert_matches = { version = "1.5.0" }
 hex = { version = "0.4.3" }
 jsonschema = { version = "0.16.1" }
 papyrus_storage = { path = "../papyrus_storage", features = ["testing"] }
+test_utils = { path = "../test_utils" }
 starknet_api = { git = "https://github.com/starkware-libs/starknet-api", features = ["testing"] , rev = "f75c4e2" }
 web3 = { version = "0.18.0" }
