+++ crates/papyrus_storage/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [features]
 default = ["starknet_api"]
-testing = ["tempfile", "starknet_api/testing", "rand"]
+testing = ["tempfile", "starknet_api/testing", "rand", "test_utils"]
 
 [dependencies]
 bincode = { version = "1.3.3" }
@@ -21,6 +21,7 @@ serde = { version = "1.0.130", features = ["derive"] }
 serde_json = { version = "1.0.81" }
 starknet_api = { git = "https://github.com/starkware-libs/starknet-api", rev = "f75c4e2", optional = true }
 tempfile = { version = "3.3.0", optional = true }
+test_utils = { path = "../test_utils", optional = true }
 thiserror = { version = "1.0.31" }
 tokio = { version = "1.18.2", features = ["full","sync"] }
 tokio-stream = { version = "0.1.8" }
