+++ crates/papyrus_storage/src/body/body_test.rs
@@ -1,10 +1,11 @@
 use assert_matches::assert_matches;
 use starknet_api::block::{BlockBody, BlockNumber};
 use starknet_api::transaction::TransactionOffsetInBlock;
+use test_utils::{get_test_block, get_test_body};
 
 use crate::body::events::ThinTransactionOutput;
 use crate::body::{BodyStorageReader, BodyStorageWriter};
-use crate::test_utils::{get_test_block, get_test_body, get_test_storage};
+use crate::test_utils::get_test_storage;
 use crate::{StorageError, StorageWriter, TransactionIndex};
 
 #[tokio::test]
