+++ crates/papyrus_storage/src/body/events_test.rs
@@ -2,11 +2,12 @@ use starknet_api::core::{ContractAddress, PatriciaKey};
 use starknet_api::hash::StarkHash;
 use starknet_api::patky;
 use starknet_api::transaction::{EventIndexInTransactionOutput, TransactionOffsetInBlock};
+use test_utils::get_test_block_with_events;
 
 use crate::body::events::EventsReader;
 use crate::body::BodyStorageWriter;
 use crate::header::HeaderStorageWriter;
-use crate::test_utils::{get_test_block_with_events, get_test_storage};
+use crate::test_utils::get_test_storage;
 use crate::{EventIndex, TransactionIndex};
 
 #[tokio::test]
