+++ crates/test_utils/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+
+[dependencies]
+indexmap = { version = "1.9.2", features = ["serde"] }
+log = { version = "0.4" }
+rand = { version = "0.8.5" }
+serde = { version = "1.0.130", features = ["derive"] }
+serde_json = { version = "1.0.81" }
+starknet_api = { git = "https://github.com/starkware-libs/starknet-api", rev = "f75c4e2" }
+
+[dev-dependencies]
